+++ src/main/java/com/lekwacious/poll/security/JwtTokenProvider.java
@@ -29,7 +29,7 @@ public class JwtTokenProvider {
         Date expiryDate =new Date(now.getTime() + jwtExpirationInMs);
 
         return Jwts.builder()
-                .setSubject(Integer.toString(userPrincipal.getId()))
+                .setSubject(Long.toString(userPrincipal.getId()))
                 .setIssuedAt(new Date())
                 .setExpiration(expiryDate)
                 .signWith(SignatureAlgorithm.HS512, jwtSecret)
